+++ NEWS.md
@@ -1,3 +1,3 @@
-# multimod (development version)
+# multimod (development version 0.0.1.9005)
 
-* Initial GitHub release.
+* Initial GitHub release (June 1, 2025).